libft

+++ libft/src/ft_bzero.c
@@ -19,7 +19,7 @@
  * @param s pointer of data
  * @param n size in byte to erase
  */
-void	ft_bzero(void *s, unsigned int n)
+void	ft_bzero(void *s, size_t n)
 {
 	unsigned int	i;
